rag/requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
langchain==0.1.0
langchain-community==0.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
Pillow==10.1.0
faiss-cpu==1.7.4
ollama==0.1.6
python-dotenv==1.0.0